+++ src/main/java/racingcar/model/Cars.java
@@ -5,7 +5,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-public class CarRepository {
+public class Cars {
     private final List<Car> cars = new ArrayList<>();
 
     public void save(Car car) {
@@ -29,20 +29,5 @@ public class CarRepository {
         return cars.stream()
                 .map(CarDto::new)
                 .collect(Collectors.toList());
-    }
-
-    public List<CarDto> getWinners() {
-        List<Car> winnerCars = cars.stream()
-                .filter(car -> car.matchPosition(getMaxPosition()))
-                .collect(Collectors.toList());
-        return carsToCarDtos(winnerCars);
-    }
-
-    private int getMaxPosition() {
-        int maxPosition = 0;
-        for (Car car : cars) {
-            maxPosition = car.getBiggerPosition(maxPosition);
-        }
-        return maxPosition;
     }
 }
