+++ src/test/java/racingcar/service/CarServiceTest.java
@@ -1,6 +1,5 @@
 package racingcar.service;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import racingcar.model.Car;
@@ -12,18 +11,12 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class CarServiceTest {
-    private CarService carService;
-
-    @BeforeEach
-    void setUp() {
-        List<String> names = Arrays.asList("범고래", "소주캉");
-        carService.createCars(names);
-    }
 
 
     @Test
     @DisplayName("CarDto를 받아서 Car객체로 반환한다.")
     void toCar() {
+        CarService carService = new CarService();
         String name = "범고래";
         CarDto carDto = new CarDto(name);
 
@@ -34,6 +27,7 @@ public class CarServiceTest {
     @Test
     @DisplayName("CarDto 리스트를 받아서 Car 리스트로 반환한다.")
     void toCars() {
+        CarService carService = new CarService();
         String name1 = "범고래";
         String name2 = "소주캉";
         List<CarDto> input = Arrays.asList(new CarDto(name1), new CarDto(name2));
